+++ CHANGES.md
@@ -47,6 +47,7 @@ Changes in 12.5
 - JS remove inline code from program in MassRequests.php & MassRequests.js
 - JS remove inline code from program in MassSchedule.php & MassSchedule.js
 - JS remove inline code from program in Requests.php & Requests.js
+- JS remove inline code from program in Schedule.php & Schedule.js
 
 Changes in 12.4.3
 -----------------